requirements.txt
fastapi[standard]
motor[srv]
passlib[bcrypt]
emails
pydantic_settings
pyjwt
